requirements.txt
llama-index-core
llama-index-vector-stores-milvus
llama-index-embeddings-huggingface
pymilvus
milvus-lite
sentence-transformers
transformers
rouge-score
sacrebleu
accelerate
bitsandbytes
pandas
requests
psutil
matplotlib
seaborn
